created new authorization controller class to use in login/signup
