+++ app/src/main/java/com/dev/moosic/AuthorizationController.kt
@@ -1,5 +1,7 @@
 package com.dev.moosic
 
+import android.content.Context
+
 interface AuthorizationController {
     val CLIENT_ID: String
         get() = "7b7fed9bf37945818d20992b055ac63b"
